docker-folder/producer/requirements.txt
praw
kafka-python
pandas==0.24.*